pyproject.toml
[tool.poetry]
name = "myoconverter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
mujoco = "^2.3.6"
loguru = "^0.7.0"
lxml = "^4.9.2"
numpy = "^1.25.0"
scipy = "^1.11.1"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.1"
trimesh = "^3.22.3"
pyvista = "^0.40.0"
networkx = "^3.1"
fpdf2 = "^2.7.4"
seaborn = "^0.12.2"


[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
